refactor(input): extract src/input.rs into par-term-input subcrate
Create new par-term-input crate containing:
- lib.rs: InputHandler for converting winit events to VT escape sequences

Only dependency fix: crate::config::OptionKeyMode -> par_term_config::OptionKeyMode.
src/input.rs replaced with thin `pub use par_term_input::*;` re-export shim.

Dependencies: par-term-config (OptionKeyMode), arboard, winit, log.

Closes #180
